rafactor: MovieStatistic not calculate percentage

+++ src/main/java/com/kronusboss/cine/statistic/adapter/controller/dto/MovieStatisticResponseDto.java
@@ -18,7 +18,7 @@ public class MovieStatisticResponseDto {
 	private Integer totalNumberOfMovies;
 	private Double averageRate;
 	private Integer averageRuntime;
-	private Map<String, Double> moviesByGender;
+	private Map<String, Integer> moviesByGender;
 	private Map<String, Long> moviesSixMonthsAgo;
 
 	public MovieStatisticResponseDto(MovieStatistic statistic) {
